refactor: table패키지 추가
- Table 관련 클래스 이동
