+++ src/main/java/domain/table/Table.java
@@ -1,0 +1,14 @@
+package domain.table;
+
+public class Table {
+	private final int number;
+
+	public Table(final int number) {
+		this.number = number;
+	}
+
+	@Override
+	public String toString() {
+		return Integer.toString(number);
+	}
+}
