+++ src/main/java/view/OutputView.java
@@ -3,7 +3,7 @@ package view;
 import java.util.List;
 
 import domain.menu.Menu;
-import domain.Table;
+import domain.table.Table;
 
 public class OutputView {
 	private static final String TOP_LINE = "┌ ─ ┐";
